tests: split the config-compiler unit tests into per-area files
codes-config-compiler-test.cxx had grown past 1800 lines. Split it into three
per-area translation units:

  - codes-config-compiler-test.cxx  -- core ConfigCompiler suite
  - codes-unit-convert-test.cxx     -- UnitConvert + ConfigCompilerUnits
  - codes-config-workload-test.cxx  -- Simulation / Workload / Jobs

Fixtures and lookup helpers used by more than one TU (find_section /
find_subsection / find_key / params_value, the kFlatBase /
kParametricDragonfly / kGroupsBase YAML fixtures, and the dragonfly_with
builder) move into a new config-compiler-test-util.h as inline
definitions; single-TU helpers stay file-local in their TU's anonymous
namespace. Pure reorganization: the same 116 tests, no renames or assertion changes.
