+++ tests/codes-config-compiler-test.cxx
@@ -5,103 +5,26 @@
 // tests assert on the returned compiled_config plain data rather than on any
 // simulator behavior. This is the first real unit-test consumer of the vendored
 // GoogleTest framework; the model-net equivalence tests cover the emit path.
+//
+// The suite is split across multiple translation units that GoogleTest links into
+// one binary; this file holds the core ConfigCompiler suite (topology sanity,
+// schema_version, malformed input, the flat / parametric / groups forms,
+// pass-through sections, reserved names, the derived-key guard, and includes).
+// Shared fixtures and lookup helpers live in config-compiler-test-util.h.
+#include "config-compiler-test-util.h"
 #include "config_compiler.h"
-#include "unit_convert.h"
 
 #include <gtest/gtest.h>
 
-#include <cstdlib>
 #include <string>
+#include <vector>
 
-using codes::config::classified_value;
-using codes::config::classify_value;
 using codes::config::compile;
 using codes::config::compiled_config;
 using codes::config::compiled_key;
 using codes::config::compiled_section;
 using codes::config::config_error;
-using codes::config::format_number;
 using codes::config::parse_includes;
-using codes::config::quantity;
-using codes::config::value_form;
-
-namespace {
-
-const compiled_section* find_section(const compiled_config& c, const char* name) {
-    for (const compiled_section& s : c.sections)
-        if (s.name == name)
-            return &s;
-    return nullptr;
-}
-
-const compiled_section* find_subsection(const compiled_section& s, const char* name) {
-    for (const compiled_section& sub : s.subsections)
-        if (sub.name == name)
-            return &sub;
-    return nullptr;
-}
-
-const compiled_key* find_key(const compiled_section& s, const char* name) {
-    for (const compiled_key& k : s.keys)
-        if (k.name == name)
-            return &k;
-    return nullptr;
-}
-
-// A minimal, valid flat-network config; tests append a `sections:` block to it.
-const char* kFlatBase = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    message_size: 464
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-)";
-
-// A minimal, valid parametric dragonfly config. With num_routers: 4 the
-// derivation is num_cn = 4/2 = 2, num_groups = 4*2+1 = 9, so repetitions =
-// num_groups*num_routers = 36, terminals per rep = 2, one router per rep.
-const char* kParametricDragonfly = R"(
-schema_version: 1
-components:
-  compute_host:
-    model: nw-lp
-topology:
-  format: parametric
-  fabric:
-    model: dragonfly
-    shape:
-      num_routers: 4
-  hosts:
-    component: compute_host
-)";
-
-// An explicit LP-groups config: two groups, custom LP types, an annotation, and
-// a PARAMS block -- the layout escape hatch for non-network configs.
-const char* kGroupsBase = R"(
-schema_version: 1
-topology:
-  format: groups
-  params:
-    message_size: 256
-  groups:
-    GRP1:
-      repetitions: 2
-      lps:
-        a: 1
-        b: 2
-        a@foo: 1
-    GRP2:
-      repetitions: 3
-      lps:
-        c: 2
-)";
-
-} // namespace
 
 // --- topology sanity (the compiler still does its normal job) ---------------
 
@@ -941,885 +864,4 @@ TEST(ConfigCompiler, ParseIncludesExtractsList) {
     ASSERT_EQ(many.size(), 2u);
     EXPECT_EQ(many.at(0), "a.yaml");
     EXPECT_EQ(many.at(1), "b.yaml");
-}
-
-// ============================================================================
-// Unit-bearing values
-// ============================================================================
-//
-// A dimensioned config value may be a bare number in the model's internal unit
-// or a unit-bearing string that is converted to that unit. The parsing/
-// conversion core (unit_convert) is tested directly first, then end-to-end
-// through compile().
-
-namespace {
-
-constexpr double KIB = 1024.0;
-constexpr double MIB = 1024.0 * 1024.0;
-constexpr double GIB = 1024.0 * 1024.0 * 1024.0;
-
-// Fetch the first value of a PARAMS key from a compiled config (empty if absent).
-std::string params_value(const compiled_config& c, const char* key) {
-    const compiled_section* p = find_section(c, "PARAMS");
-    if (!p)
-        return {};
-    const compiled_key* k = find_key(*p, key);
-    return (k && !k->values.empty()) ? k->values.at(0) : std::string();
-}
-
-// A parametric dragonfly (num_routers: 4) carrying one extra fabric-level scalar
-// key -- the pass-through path most model knobs take into PARAMS.
-std::string dragonfly_with(const std::string& key, const std::string& value) {
-    return std::string(R"(
-schema_version: 1
-components:
-  compute_host: { model: nw-lp }
-topology:
-  format: parametric
-  fabric:
-    model: dragonfly
-    shape:
-      num_routers: 4
-    )") + key +
-           ": " + value + R"(
-  hosts:
-    component: compute_host
-)";
-}
-
-// A single-key parametric fabric of the named model, for the flat-bandwidth
-// models (torus/fattree) whose link_bandwidth unit differs.
-std::string torus_link_bw(const std::string& value) {
-    return std::string(R"(
-schema_version: 1
-components:
-  compute_host: { model: nw-lp }
-topology:
-  format: parametric
-  fabric:
-    model: torus
-    shape:
-      n_dims: 1
-      dim_length: "8"
-    link_bandwidth: )") +
-           value + R"(
-  hosts:
-    component: compute_host
-)";
-}
-
-std::string fattree_link_bw(const std::string& value) {
-    return std::string(R"(
-schema_version: 1
-components:
-  compute_host: { model: nw-lp }
-topology:
-  format: parametric
-  fabric:
-    model: fattree
-    shape:
-      num_levels: 3
-      switch_count: 32
-      switch_radix: 8
-    link_bandwidth: )") +
-           value + R"(
-  hosts:
-    component: compute_host
-)";
-}
-
-// A flat simplenet component carrying one extra scalar key.
-std::string simplenet_with(const std::string& key, const std::string& value) {
-    return std::string(R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    )") + key +
-           ": " + value + R"(
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-)";
-}
-
-} // namespace
-
-// --- classify_value: bare / plain / units --------------------------------
-
-TEST(UnitConvert, ClassifiesBareNumberAndPlainString) {
-    EXPECT_EQ(classify_value("512").form, value_form::bare_number);
-    EXPECT_EQ(classify_value("5.25").form, value_form::bare_number);
-    EXPECT_EQ(classify_value("1.5e6").form, value_form::bare_number); // scientific = bare
-    EXPECT_EQ(classify_value("-3").form, value_form::bare_number);
-
-    EXPECT_EQ(classify_value("adaptive").form, value_form::plain);
-    EXPECT_EQ(classify_value("some/path.conf").form, value_form::plain);
-    EXPECT_EQ(classify_value("").form, value_form::plain);
-    EXPECT_EQ(classify_value("4,2,2").form, value_form::unknown_suffix); // number + junk
-}
-
-TEST(UnitConvert, ParsesTimeSuffixesToNanoseconds) {
-    classified_value ns = classify_value("40ns");
-    EXPECT_EQ(ns.form, value_form::with_unit);
-    EXPECT_EQ(ns.kind, quantity::time);
-    EXPECT_DOUBLE_EQ(ns.canonical, 40.0);
-    EXPECT_DOUBLE_EQ(classify_value("10us").canonical, 10000.0);
-    EXPECT_DOUBLE_EQ(classify_value("1.5ms").canonical, 1.5e6);
-    EXPECT_DOUBLE_EQ(classify_value("2s").canonical, 2.0e9);
-}
-
-TEST(UnitConvert, ParsesSizeSuffixesToBytes) {
-    classified_value b = classify_value("1500B");
-    EXPECT_EQ(b.kind, quantity::size);
-    EXPECT_DOUBLE_EQ(b.canonical, 1500.0);
-    EXPECT_DOUBLE_EQ(classify_value("2KiB").canonical, 2048.0);
-    EXPECT_DOUBLE_EQ(classify_value("4MiB").canonical, 4.0 * MIB);
-    EXPECT_DOUBLE_EQ(classify_value("1GiB").canonical, GIB);
-    // decimal SI forms are accepted and are 1000-based.
-    EXPECT_DOUBLE_EQ(classify_value("2KB").canonical, 2000.0);
-    EXPECT_DOUBLE_EQ(classify_value("1MB").canonical, 1.0e6);
-}
-
-TEST(UnitConvert, ParsesBandwidthBitAndByteRates) {
-    // bit rates are decimal, /8 to bytes/second.
-    classified_value g = classify_value("100Gbps");
-    EXPECT_EQ(g.kind, quantity::bandwidth);
-    EXPECT_DOUBLE_EQ(g.canonical, 100.0e9 / 8.0);
-    EXPECT_DOUBLE_EQ(classify_value("8bps").canonical, 1.0);
-    // byte rates: decimal and binary forms.
-    EXPECT_DOUBLE_EQ(classify_value("2.5GBps").canonical, 2.5e9);
-    EXPECT_DOUBLE_EQ(classify_value("1GiBps").canonical, GIB);
-    EXPECT_DOUBLE_EQ(classify_value("1KiBps").canonical, KIB);
-}
-
-TEST(UnitConvert, SuffixMatchIsCaseSensitiveBitVsByte) {
-    // lowercase 'b' = bits, uppercase 'B' = bytes: the two differ by a factor 8.
-    EXPECT_DOUBLE_EQ(classify_value("1Gbps").canonical, 1.0e9 / 8.0);
-    EXPECT_DOUBLE_EQ(classify_value("1GBps").canonical, 1.0e9);
-}
-
-TEST(UnitConvert, RejectsAndPassesTrailingText) {
-    EXPECT_EQ(classify_value("5xz").form, value_form::unknown_suffix);
-    EXPECT_EQ(classify_value("512qux").form, value_form::unknown_suffix);
-    // surrounding whitespace is ignored.
-    EXPECT_EQ(classify_value("  5ms  ").form, value_form::with_unit);
-    EXPECT_DOUBLE_EQ(classify_value("  5ms  ").canonical, 5.0e6);
-}
-
-// --- format_number: integers vs fractions, round-trip --------------------
-
-TEST(UnitConvert, FormatsIntegerValuesWithoutDecimalPoint) {
-    EXPECT_EQ(format_number(2048.0), "2048");
-    EXPECT_EQ(format_number(1500.0), "1500");
-    EXPECT_EQ(format_number(0.0), "0");
-    EXPECT_EQ(format_number(1.0e9), "1000000000");
-}
-
-TEST(UnitConvert, FormatsFractionsAsPlainDecimalThatRoundTrips) {
-    EXPECT_EQ(format_number(1.5), "1.5");
-    EXPECT_EQ(format_number(2.5), "2.5");
-    // a converted bandwidth: 2.5 GB/s in GiB/s -- no exponent, round-trips.
-    double v = 2.5e9 / GIB;
-    std::string s = format_number(v);
-    EXPECT_EQ(s.find('e'), std::string::npos);
-    EXPECT_EQ(s.find('E'), std::string::npos);
-    EXPECT_DOUBLE_EQ(std::strtod(s.c_str(), nullptr), v);
-}
-
-// --- end-to-end: conversion through compile() ----------------------------
-
-TEST(ConfigCompilerUnits, ConvertsSizeSuffixToBytes) {
-    EXPECT_EQ(params_value(compile(dragonfly_with("packet_size", "2KiB")), "packet_size"), "2048");
-    EXPECT_EQ(params_value(compile(dragonfly_with("message_size", "1500B")), "message_size"),
-              "1500");
-    EXPECT_EQ(params_value(compile(dragonfly_with("chunk_size", "4MiB")), "chunk_size"),
-              format_number(4.0 * MIB));
-}
-
-TEST(ConfigCompilerUnits, ConvertsTimeSuffixToNanoseconds) {
-    EXPECT_EQ(params_value(compile(dragonfly_with("router_delay", "1.5us")), "router_delay"),
-              "1500");
-    EXPECT_EQ(params_value(compile(simplenet_with("net_startup_ns", "5ms")), "net_startup_ns"),
-              "5000000");
-}
-
-TEST(ConfigCompilerUnits, ConvertsBandwidthToGiBPerSecond) {
-    // cn_bandwidth is read in GiB/s; "2.5GBps" is 2.5 GB/s (decimal) = 2.5e9 B/s.
-    std::string got =
-        params_value(compile(dragonfly_with("cn_bandwidth", "2.5GBps")), "cn_bandwidth");
-    EXPECT_EQ(got, format_number(2.5e9 / GIB));
-}
-
-TEST(ConfigCompilerUnits, BandwidthConversionIsPerModel) {
-    // The same "8Gbps" (1e9 B/s) resolves differently: torus reads GiB/s, fattree
-    // reads bytes/ns (= GB/s decimal). fattree lands on a clean 1.
-    std::string torus = params_value(compile(torus_link_bw("8Gbps")), "link_bandwidth");
-    std::string fattree = params_value(compile(fattree_link_bw("8Gbps")), "link_bandwidth");
-    EXPECT_EQ(fattree, "1");
-    EXPECT_EQ(torus, format_number(1.0e9 / GIB));
-    EXPECT_NE(torus, fattree);
-}
-
-TEST(ConfigCompilerUnits, ConvertsSimplenetBandwidthToMebibytesPerSecond) {
-    // net_bw_mbps is read in MiB/s despite the name; 1 GiB/s = 1024 MiB/s.
-    EXPECT_EQ(params_value(compile(simplenet_with("net_bw_mbps", "1GiBps")), "net_bw_mbps"),
-              "1024");
-}
-
-TEST(ConfigCompilerUnits, ConvertsMicrosecondCountingWindow) {
-    // dragonfly QoS counting_* windows are read in microseconds, not ns.
-    EXPECT_EQ(params_value(compile(R"(
-schema_version: 1
-components:
-  compute_host: { model: nw-lp }
-topology:
-  format: parametric
-  fabric:
-    model: dragonfly-dally
-    shape:
-      num_routers: 4
-      num_groups: 9
-      num_cns_per_router: 2
-      num_global_channels: 2
-    counting_start: 5ms
-  hosts:
-    component: compute_host
-)"),
-                           "counting_start"),
-              "5000"); // 5 ms = 5000 us
-}
-
-TEST(ConfigCompilerUnits, ConvertsLinkClassBandwidthBlock) {
-    // the links: sugar reaches the same conversion (local_bandwidth here).
-    std::string got = params_value(compile(R"(
-schema_version: 1
-components:
-  compute_host: { model: nw-lp }
-topology:
-  format: parametric
-  fabric:
-    model: dragonfly
-    shape:
-      num_routers: 4
-    links:
-      local: { bandwidth: 1GiBps }
-  hosts:
-    component: compute_host
-)"),
-                                   "local_bandwidth");
-    EXPECT_EQ(got, "1"); // 1 GiB/s in GiB/s
-}
-
-// --- bare numbers keep their exact spelling (identity) --------------------
-
-TEST(ConfigCompilerUnits, BareNumbersPassThroughVerbatim) {
-    // A bare number already means the model's internal unit and is emitted
-    // byte-for-byte -- this is what keeps every existing .conf twin equivalent.
-    EXPECT_EQ(params_value(compile(dragonfly_with("packet_size", "512")), "packet_size"), "512");
-    EXPECT_EQ(params_value(compile(dragonfly_with("cn_bandwidth", "5.25")), "cn_bandwidth"),
-              "5.25");
-    EXPECT_EQ(params_value(compile(simplenet_with("net_bw_mbps", "20000")), "net_bw_mbps"),
-              "20000");
-    // a non-unit string on an unclassified knob passes through untouched.
-    const compiled_config c = compile(torus_link_bw("2.0"));
-    EXPECT_EQ(params_value(c, "dim_length"), "8"); // number+junk on unclassified: verbatim
-}
-
-// --- negative suite -------------------------------------------------------
-
-TEST(ConfigCompilerUnits, RejectsUnknownUnitSuffixOnDimensionedParam) {
-    EXPECT_THROW(compile(dragonfly_with("packet_size", "512qux")), config_error);
-    EXPECT_THROW(compile(dragonfly_with("chunk_size", "32xyz")), config_error);
-}
-
-TEST(ConfigCompilerUnits, RejectsWrongQuantityUnit) {
-    // a time unit on a size parameter is a mistake, not a silent misread.
-    EXPECT_THROW(compile(dragonfly_with("packet_size", "5ms")), config_error);
-    // a size unit on a bandwidth parameter, likewise.
-    EXPECT_THROW(compile(dragonfly_with("cn_bandwidth", "5MiB")), config_error);
-}
-
-TEST(ConfigCompilerUnits, RejectsNegativeDimensionedValue) {
-    EXPECT_THROW(compile(dragonfly_with("cn_bandwidth", "-1Gbps")), config_error);
-    EXPECT_THROW(compile(dragonfly_with("packet_size", "-5B")), config_error);
-}
-
-TEST(ConfigCompilerUnits, RejectsUnitSuffixOnUnclassifiableParam) {
-    // num_vcs is a plain count the model atof()s; a unit there would be silently
-    // truncated to its bare number, so it is rejected up front.
-    EXPECT_THROW(compile(dragonfly_with("num_vcs", "100Gbps")), config_error);
-    EXPECT_THROW(compile(dragonfly_with("num_vcs", "4KiB")), config_error);
-}
-
-// ---------------------------------------------------------------------------
-// simulation: block -- run-level settings (end_time, pe_mem_factor) resolved to
-// the PARAMS keys codes_mapping reads.
-// ---------------------------------------------------------------------------
-
-namespace {
-// kFlatBase plus a trailing top-level simulation: block (body already indented).
-std::string flat_with_simulation(const std::string& body) {
-    return std::string(kFlatBase) + "simulation:\n" + body;
-}
-} // namespace
-
-TEST(ConfigCompilerSimulation, EmitsEndTimeAndMemFactorToParams) {
-    // both settings land in PARAMS -- the exact section/keys codes_mapping reads.
-    compiled_config c = compile(flat_with_simulation("  end_time: 5000\n  pe_mem_factor: 512\n"));
-    EXPECT_EQ(params_value(c, "end_time"), "5000");
-    EXPECT_EQ(params_value(c, "pe_mem_factor"), "512");
-}
-
-TEST(ConfigCompilerSimulation, EndTimeAcceptsTimeUnits) {
-    // a bare number is nanoseconds; a unit-bearing value converts to ns.
-    EXPECT_EQ(params_value(compile(flat_with_simulation("  end_time: 100us\n")), "end_time"),
-              "100000");
-    EXPECT_EQ(params_value(compile(flat_with_simulation("  end_time: 2ms\n")), "end_time"),
-              "2000000");
-    EXPECT_EQ(params_value(compile(flat_with_simulation("  end_time: 250\n")), "end_time"), "250");
-}
-
-TEST(ConfigCompilerSimulation, RejectsUnknownKeyInBlock) {
-    EXPECT_THROW(compile(flat_with_simulation("  bogus: 1\n")), config_error);
-}
-
-TEST(ConfigCompilerSimulation, RejectsNonPositiveMemFactor) {
-    EXPECT_THROW(compile(flat_with_simulation("  pe_mem_factor: 0\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  pe_mem_factor: -4\n")), config_error);
-}
-
-TEST(ConfigCompilerSimulation, RejectsNonIntegerMemFactor) {
-    EXPECT_THROW(compile(flat_with_simulation("  pe_mem_factor: 3.5\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  pe_mem_factor: abc\n")), config_error);
-}
-
-TEST(ConfigCompilerSimulation, RejectsMalformedEndTime) {
-    // not a number, not positive, negative, and an unrecognized suffix.
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: abc\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: 0\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: -5us\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: 100xyz\n")), config_error);
-}
-
-TEST(ConfigCompilerSimulation, RejectsWrongQuantityUnitOnEndTime) {
-    // a size or bandwidth unit on a time value is a mistake, not a silent misread.
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: 2KiB\n")), config_error);
-    EXPECT_THROW(compile(flat_with_simulation("  end_time: 5Gbps\n")), config_error);
-}
-
-TEST(ConfigCompilerSimulation, ConflictsWithPassthroughParam) {
-    // A key set both as a fabric pass-through and in simulation: would be silently
-    // dropped by the config store's first-match lookup, so it is rejected.
-    EXPECT_THROW(compile(dragonfly_with("pe_mem_factor", "256") +
-                         "simulation:\n  pe_mem_factor: 512\n"),
-                 config_error);
-    EXPECT_THROW(compile(dragonfly_with("end_time", "5000") + "simulation:\n  end_time: 9000\n"),
-                 config_error);
-}
-
-TEST(ConfigCompilerSimulation, PassthroughParamAloneStillWorks) {
-    // Without a simulation: block, pe_mem_factor on the fabric passes through to
-    // PARAMS exactly as before -- no regression for existing configs.
-    compiled_config c = compile(dragonfly_with("pe_mem_factor", "256"));
-    EXPECT_EQ(params_value(c, "pe_mem_factor"), "256");
-}
-
-TEST(ConfigCompilerSimulation, WorksWithParametricTopology) {
-    compiled_config c =
-        compile(std::string(kParametricDragonfly) + "simulation:\n  end_time: 50us\n");
-    EXPECT_EQ(params_value(c, "end_time"), "50000");
-}
-
-TEST(ConfigCompilerSimulation, WorksWithExplicitGroups) {
-    // the explicit-groups form builds PARAMS itself; the simulation keys still
-    // land in it.
-    compiled_config c = compile(std::string(kGroupsBase) + "simulation:\n  pe_mem_factor: 64\n");
-    EXPECT_EQ(params_value(c, "pe_mem_factor"), "64");
-}
-
-TEST(ConfigCompilerSimulation, ConflictsWithExplicitGroupsParam) {
-    // a pe_mem_factor written directly in the explicit-groups params: collides
-    // with simulation.pe_mem_factor exactly like the pass-through fabric case.
-    const char* yaml = R"(
-schema_version: 1
-topology:
-  format: groups
-  params:
-    pe_mem_factor: 256
-  groups:
-    GRP:
-      repetitions: 1
-      lps:
-        a: 1
-simulation:
-  pe_mem_factor: 512
-)";
-    EXPECT_THROW(compile(yaml), config_error);
-}
-
-// --- workload: shortcut and jobs: block -----------------------------------
-//
-// A node's workload (what it runs) is configured separately from the node model
-// (what it is): an inline `workload:` on the topology's compute component is the
-// single-workload shortcut, and a top-level `jobs:` block is the explicit
-// multi-job form. Both lower to a WORKLOAD (or JOBS) section a synthetic main
-// reads. Only synthetic is wired; other types are recognized but rejected.
-
-namespace {
-
-// Fetch the first value of a WORKLOAD-section key (empty if absent).
-std::string workload_value(const compiled_config& c, const char* key) {
-    const compiled_section* w = find_section(c, "WORKLOAD");
-    if (!w)
-        return {};
-    const compiled_key* k = find_key(*w, key);
-    return (k && !k->values.empty()) ? k->values.at(0) : std::string();
-}
-
-// Deep structural equality of two compiled configs (section/key/value/subsection,
-// order included) -- used to prove two spellings compile to the identical tree.
-bool key_eq(const compiled_key& a, const compiled_key& b) {
-    return a.name == b.name && a.values == b.values;
-}
-bool section_eq(const compiled_section& a, const compiled_section& b) {
-    if (a.name != b.name || a.keys.size() != b.keys.size() ||
-        a.subsections.size() != b.subsections.size())
-        return false;
-    for (size_t i = 0; i < a.keys.size(); ++i)
-        if (!key_eq(a.keys[i], b.keys[i]))
-            return false;
-    for (size_t i = 0; i < a.subsections.size(); ++i)
-        if (!section_eq(a.subsections[i], b.subsections[i]))
-            return false;
-    return true;
-}
-bool config_eq(const compiled_config& a, const compiled_config& b) {
-    if (a.sections.size() != b.sections.size())
-        return false;
-    for (size_t i = 0; i < a.sections.size(); ++i)
-        if (!section_eq(a.sections[i], b.sections[i]))
-            return false;
-    return true;
-}
-
-// A flat 4-node config whose compute component carries an inline workload: block.
-// `body` supplies the workload's keys, each indented six spaces.
-std::string flat_with_workload(const std::string& body) {
-    return std::string(R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    message_size: 464
-    workload:
-)") + body +
-           R"(
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-)";
-}
-
-// A flat 4-node config with a top-level jobs: block. `body` supplies the job
-// list entries (each `- id:` indented two spaces).
-std::string flat_with_jobs(const std::string& body) {
-    return std::string(kFlatBase) + "jobs:\n" + body;
-}
-
-} // namespace
-
-TEST(ConfigCompilerWorkload, InlineEmitsResolvedWorkloadSection) {
-    // payload_size resolves to bytes, arrival_time to nanoseconds; traffic stays
-    // a verbatim pattern name and num_messages a positive integer.
-    compiled_config c = compile(flat_with_workload("      type: synthetic\n"
-                                                   "      traffic: uniform\n"
-                                                   "      num_messages: 30\n"
-                                                   "      payload_size: \"2KiB\"\n"
-                                                   "      arrival_time: \"1us\"\n"));
-    EXPECT_EQ(workload_value(c, "type"), "synthetic");
-    EXPECT_EQ(workload_value(c, "traffic"), "uniform");
-    EXPECT_EQ(workload_value(c, "num_messages"), "30");
-    EXPECT_EQ(workload_value(c, "payload_size"), "2048");
-    EXPECT_EQ(workload_value(c, "arrival_time"), "1000");
-}
-
-TEST(ConfigCompilerWorkload, NoWorkloadEmitsNoSection) {
-    // A config without workload:/jobs: emits no WORKLOAD section, so a legacy
-    // config's tree is unchanged.
-    compiled_config c = compile(kFlatBase);
-    EXPECT_EQ(find_section(c, "WORKLOAD"), nullptr);
-    EXPECT_EQ(find_section(c, "JOBS"), nullptr);
-}
-
-TEST(ConfigCompilerWorkload, RejectsUnknownKey) {
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n      bogus: 1\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsMissingType) {
-    EXPECT_THROW(compile(flat_with_workload("      traffic: uniform\n")), config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsNonSyntheticType) {
-    // dumpi etc. are recognized as a workload shape but not yet configurable here.
-    EXPECT_THROW(compile(flat_with_workload("      type: dumpi\n      trace: app.dumpi\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsWrongUnitOnArrivalTime) {
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n"
-                                            "      arrival_time: \"2KiB\"\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsWrongUnitOnPayloadSize) {
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n"
-                                            "      payload_size: \"1us\"\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsNonPositiveMessageCount) {
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n      num_messages: 0\n")),
-                 config_error);
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n      num_messages: -5\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsNonIntegerMessageCount) {
-    EXPECT_THROW(compile(flat_with_workload("      type: synthetic\n      num_messages: 3.5\n")),
-                 config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsWorkloadOnNonComputeComponent) {
-    // a workload: on a component the topology does not run is meaningless; reject
-    // it rather than silently ignore it.
-    const char* yaml = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-  other:
-    model: nw-lp
-    workload:
-      type: synthetic
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-)";
-    EXPECT_THROW(compile(yaml), config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsWorkloadWithExplicitGroups) {
-    const char* yaml = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    workload:
-      type: synthetic
-topology:
-  format: groups
-  groups:
-    GRP:
-      repetitions: 1
-      lps:
-        a: 1
-)";
-    EXPECT_THROW(compile(yaml), config_error);
-}
-
-TEST(ConfigCompilerWorkload, RejectsInlineWorkloadAndJobsTogether) {
-    // exactly one form is allowed.
-    const char* yaml = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    workload:
-      type: synthetic
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-jobs:
-  - id: j
-    workload:
-      type: synthetic
-    ranks: 4
-    placement:
-      policy: contiguous
-)";
-    EXPECT_THROW(compile(yaml), config_error);
-}
-
-TEST(ConfigCompilerJobs, SingleAllNodesJobDesugarsToInlineWorkload) {
-    // A single synthetic job placed contiguously on every node must compile to
-    // the exact same tree as the inline workload: shortcut.
-    const char* inln = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    message_size: 464
-    workload:
-      type: synthetic
-      traffic: uniform
-      num_messages: 30
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-)";
-    const char* jobs = R"(
-schema_version: 1
-components:
-  cn:
-    model: nw-lp
-    network: simplenet
-    message_size: 464
-topology:
-  format: flat
-  component: cn
-  nodes: 4
-jobs:
-  - id: everything
-    workload:
-      type: synthetic
-      traffic: uniform
-      num_messages: 30
-    ranks: 4
-    placement:
-      policy: contiguous
-)";
-    compiled_config a = compile(inln);
-    compiled_config b = compile(jobs);
-    EXPECT_TRUE(config_eq(a, b));
-    // and it really is the desugared shortcut: a WORKLOAD section, no JOBS.
-    EXPECT_NE(find_section(b, "WORKLOAD"), nullptr);
-    EXPECT_EQ(find_section(b, "JOBS"), nullptr);
-}
-
-TEST(ConfigCompilerJobs, MultiJobEmitsJobsSectionWithNodeAllocation) {
-    // two explicitly-placed jobs on a 4-node topology -> a JOBS section, one
-    // subsection per job carrying its workload, ranks, and resolved node list.
-    compiled_config c = compile(flat_with_jobs(R"(  - id: front
-    workload:
-      type: synthetic
-      traffic: uniform
-    ranks: 2
-    placement:
-      nodes: [0, 1]
-  - id: back
-    workload:
-      type: synthetic
-      traffic: nearest_neighbor
-    ranks: 2
-    placement:
-      nodes: [2, 3]
-)"));
-    EXPECT_EQ(find_section(c, "WORKLOAD"), nullptr);
-    const compiled_section* jobs = find_section(c, "JOBS");
-    ASSERT_NE(jobs, nullptr);
-    EXPECT_EQ(find_key(*jobs, "num_jobs")->values.at(0), "2");
-    const compiled_section* front = find_subsection(*jobs, "front");
-    const compiled_section* back = find_subsection(*jobs, "back");
-    ASSERT_NE(front, nullptr);
-    ASSERT_NE(back, nullptr);
-    EXPECT_EQ(find_key(*front, "traffic")->values.at(0), "uniform");
-    EXPECT_EQ(find_key(*front, "ranks")->values.at(0), "2");
-    const compiled_key* fnodes = find_key(*front, "nodes");
-    ASSERT_NE(fnodes, nullptr);
-    EXPECT_EQ(fnodes->values, (std::vector<std::string>{"0", "1"}));
-    const compiled_key* bnodes = find_key(*back, "nodes");
-    ASSERT_NE(bnodes, nullptr);
-    EXPECT_EQ(bnodes->values, (std::vector<std::string>{"2", "3"}));
-}
-
-TEST(ConfigCompilerJobs, ContiguousPolicyPacksInDeclaredOrder) {
-    // two contiguous jobs pack from node 0 in order: [0,1] then [2,3].
-    compiled_config c = compile(flat_with_jobs(R"(  - id: a
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { policy: contiguous }
-  - id: b
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { policy: contiguous }
-)"));
-    const compiled_section* jobs = find_section(c, "JOBS");
-    ASSERT_NE(jobs, nullptr);
-    EXPECT_EQ(find_key(*find_subsection(*jobs, "a"), "nodes")->values,
-              (std::vector<std::string>{"0", "1"}));
-    EXPECT_EQ(find_key(*find_subsection(*jobs, "b"), "nodes")->values,
-              (std::vector<std::string>{"2", "3"}));
-}
-
-TEST(ConfigCompilerJobs, RejectsDuplicateId) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: dup
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { policy: contiguous }
-  - id: dup
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsEmptyId) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: ""
-    workload: { type: synthetic }
-    ranks: 4
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsMissingWorkload) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    ranks: 4
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsNonPositiveRanks) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 0
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsNonIntegerRanks) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: two
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsMissingPlacement) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsPlacementWithBothForms) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-    placement:
-      policy: contiguous
-      nodes: [0, 1, 2, 3]
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsPlacementWithNeitherForm) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-    placement: {}
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsUnknownPlacementPolicy) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-    placement: { policy: scattered }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsNodeOutOfRange) {
-    // node 9 does not exist in a 4-node topology.
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 1
-    placement: { nodes: [9] }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsDoubleBookedNode) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: a
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { nodes: [0, 1] }
-  - id: b
-    workload: { type: synthetic }
-    ranks: 2
-    placement: { nodes: [1, 2] }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsRanksNodeListMismatch) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 3
-    placement: { nodes: [0, 1] }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsTotalRanksExceedingSlots) {
-    // two contiguous jobs of 3 ranks each need 6 slots; the topology has 4.
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: a
-    workload: { type: synthetic }
-    ranks: 3
-    placement: { policy: contiguous }
-  - id: b
-    workload: { type: synthetic }
-    ranks: 3
-    placement: { policy: contiguous }
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsQosKey) {
-    // qos is recognized but not yet wired.
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-    placement: { policy: contiguous }
-    qos: 1
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsUnknownJobKey) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: synthetic }
-    ranks: 4
-    placement: { policy: contiguous }
-    bogus: 1
-)")),
-                 config_error);
-}
-
-TEST(ConfigCompilerJobs, RejectsNonSyntheticJobWorkload) {
-    EXPECT_THROW(compile(flat_with_jobs(R"(  - id: j
-    workload: { type: dumpi, trace: app.dumpi }
-    ranks: 4
-    placement: { policy: contiguous }
-)")),
-                 config_error);
 }
